feat: mean reversion alert, TP/SL monitor, optimized indicators for 15min
- Fix Q key not responding during retry loops (price unavailable, Binance error)
- Fix execute_hotkey discarding error messages from failed buys
- Add PRICE_ALERT_ENABLED and SIGNAL_ENABLED toggle variables
- Add PRICE_BEAT_ALERT with MID phase + token cheap condition
- Add Mean Reversion alert (RSI extreme + Bollinger touch + MID phase)
- Add position TP/SL monitor with beep alerts
- Optimize indicator settings for 15min mean reversion strategy:
  RSI=5, MACD=12/26/9, BB=10/1.5, ADX=14
- Rebalance signal weights: divergence + bollinger weight increased
- Fix BB_STD parsed as int instead of float

+++ src/binance_api.py
@@ -26,7 +26,7 @@ MACD_FAST = int(os.getenv('MACD_FAST', '5'))
 MACD_SLOW = int(os.getenv('MACD_SLOW', '10'))
 MACD_SIGNAL = int(os.getenv('MACD_SIGNAL', '4'))
 BB_PERIOD = int(os.getenv('BB_PERIOD', '14'))
-BB_STD = int(os.getenv('BB_STD', '2'))
+BB_STD = float(os.getenv('BB_STD', '2'))
 ADX_PERIOD = int(os.getenv('ADX_PERIOD', '7'))
